refactor(installer): split installer into modules and add dry-run plan
