+++ src/bin/sempal-installer/download.rs
@@ -1,0 +1,3 @@
+pub(crate) fn ensure_downloads() -> Result<(), String> {
+    Ok(())
+}
